requirements.txt
ultralytics
fastapi
uvicorn
opencv-python-headless
numpy
Pillow
python-multipart